did more linting, refactored ROS Node system
